competitor_analysis-main/scrappers/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==2025.4.4
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-monorepo @ git+https://github.com/langchain-ai/langchain.git@893942651bd3278c632b37301452b027cc1e2e71
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.3.25
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.25
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.3.2
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
numpy==2.2.4
openai==1.70.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
playwright==1.51.0
pluggy==1.5.0
primp==0.14.0
propcache==0.3.1
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyee==12.1.1
Pygments==2.19.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.1
tenacity==9.1.2
text-unidecode==1.3
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
validators==0.34.0
watchfiles==1.0.4
websockets==15.0.1
xxhash==3.5.0
yarl==1.19.0
zstandard==0.23.0
